+++ frontend/src/main/resources/webapp/src/app/main/books/books-home/book-grid/book-grid.component.ts
@@ -4,7 +4,8 @@ import {MyListService} from "../../../../shared/services/my-list.service";
 @Component({
   selector: 'app-book-grid',
   templateUrl: './book-grid.component.html',
-  styleUrls: ['./book-grid.component.scss']
+  styleUrls: ['./book-grid.component.scss'],
+  providers: [MyListService]
 })
 export class BookGridComponent implements OnInit {